Transform computeChecksums in AWS.Config
