+++ src/transforms/v2-to-v3/__fixtures__/config/computeChecksums.input.js
@@ -1,0 +1,5 @@
+import AWS from "aws-sdk";
+
+const config = new AWS.Config({
+  computeChecksums: true
+});
