+++ src/transforms/v2-to-v3/__fixtures__/config/computeChecksums.output.js
@@ -1,0 +1,5 @@
+const config = {
+  // The key computeChecksums is no longer supported in v3, and can be removed.
+  // @deprecated Applicable commands of S3 will automatically compute the MD5 checksums.
+  computeChecksums: true
+};
