+++ src/transforms/v2-to-v3/config/AWS_CONFIG_KEY_MAP.ts
@@ -8,5 +8,8 @@ export interface AwsConfigKeyStatus {
  * Maps the AWS config keys with their equivalent replacement in v3.
  */
 export const AWS_CONFIG_KEY_MAP: Record<string, AwsConfigKeyStatus> = {
+  computeChecksums: {
+    deprecationMessage: "Applicable commands of S3 will automatically compute the MD5 checksums.",
+  },
   region: {},
 };
